# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignmentMutation": {
        "additionalProperties": false,
        "properties": {
          "assignedUserId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigneduserid"
          },
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          }
        },
        "required": [
          "assignedUserId",
          "expectedVersion"
        ],
        "title": "AssignmentMutation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InquiryPageResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InquirySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "InquiryPageResponse",
        "type": "object"
      },
...
      "StatusMutation": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "status": {
            "enum": [
              "new",
              "contacted",
              "qualified",
              "viewing",
              "won",
              "lost",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "expectedVersion"
        ],
        "title": "StatusMutation",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Real-Estate Lead Capture CRM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inquiries": {
      "get": {
        "operationId": "listInquiries",
        "parameters": [
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "contacted",
                    "qualified",
                    "viewing",
                    "won",
                    "lost",
                    "closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "propertyId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Propertyid"
            }
          },
          {
            "in": "query",
            "name": "assignedUserId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigneduserid"
            }
          },
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryPageResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Role denied",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Inquiries",
        "tags": [
          "crm"
        ]
      }
    },
    "/api/v1/inquiries/{inquiry_id}": {
      "get": {
        "operationId": "getInquiry",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryDetail"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Role denied",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Inquiry not found",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Inquiry",
        "tags": [
          "crm"
        ]
      }
    },
    "/api/v1/inquiries/{inquiry_id}/assignment": {
      "patch": {
        "operationId": "updateInquiryAssignment",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentMutation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryDetail"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Role or CSRF denied",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Inquiry not found",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Stale inquiry version",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Update Assignment",
        "tags": [
          "crm"
        ]
      }
    },
    "/api/v1/inquiries/{inquiry_id}/status": {
      "patch": {
        "operationId": "updateInquiryStatus",
        "parameters": [
          {
            "in": "path",
            "name": "inquiry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inquiry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusMutation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InquiryDetail"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Role or CSRF denied",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Inquiry not found",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Stale inquiry version",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Update Status",
        "tags": [
          "crm"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "listAssignees",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssigneeChoice"
                  },
                  "title": "Response Listassignees",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Role denied",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Users",
        "tags": [
          "crm"
        ]
      }
    }
  }
}